Update Image Props
