+++ src/types/image/image.types.ts
@@ -1,0 +1,8 @@
+import { ImgHTMLAttributes } from 'react';
+
+export interface ImageProps extends ImgHTMLAttributes<HTMLImageElement> {
+	src: string;
+	alt: string;
+	layout?: 'fill' | 'fixed' | 'intrinsic' | 'responsive'; // Next.js specific
+	priority?: boolean; // Next.js specific
+}
